顧客管理タブで分割
