+++ web/admin/src/types/props/customer.ts
@@ -1,0 +1,4 @@
+export interface Customer {
+  name: string
+  value: string
+}
